MAINT: Remove `plot-qualities` visualizer
The new `demux summarize` visualizer replaces this.

Fixes #59
