+++ q2_dada2/__init__.py
@@ -7,11 +7,10 @@
 # ----------------------------------------------------------------------------
 
 from ._denoise import denoise_single, denoise_paired
-from ._plot import plot_qualities
 from ._version import get_versions
 
 
 __version__ = get_versions()['version']
 del get_versions
 
-__all__ = ['denoise_single', 'denoise_paired', 'plot_qualities']
+__all__ = ['denoise_single', 'denoise_paired']
